move maps build-component function
given there will be handful of md.use() functions, we're moving the
functions within the render key's value to imported functions.
