+++ output/site.html
@@ -9,5 +9,5 @@
 <p>lorum ipsum</p>
 <h2>Subheader</h2>
 <p>lorum <a href="http://www.google.com">ipsum</a></p>
-<style>      #south-africa { height: 180px; }      </style>      <div id=south-africa></div><script>      var mymap = L.map("south-africa").setView([-33.9249, 18.4241], 12);      L.tileLayer("http://a.aerial.openstreetmap.org.za/ngi-aerial/{z}/{x}/{y}.jpg", {      attribution: 'Map data &copy; <a href="http://openstreetmap.org">OpenStreetMap</a> contributors, <a href="http://creativecommons.org/licenses/by-sa/2.0/">CC-BY-SA</a>, Imagery © <a href="http://mapbox.com">Mapbox</a>',      maxZoom: 18,      accessToken: 123456      }).addTo(mymap);      </script></body>
+<style>    #south-africa { height: 180px; }    </style>    <div id=south-africa></div><script>    var mymap = L.map("south-africa").setView([-33.9249, 18.4241], 12);    L.tileLayer("http://a.aerial.openstreetmap.org.za/ngi-aerial/{z}/{x}/{y}.jpg", {    attribution: 'Map data &copy; <a href="http://openstreetmap.org">OpenStreetMap</a> contributors, <a href="http://creativecommons.org/licenses/by-sa/2.0/">CC-BY-SA</a>, Imagery © <a href="http://mapbox.com">Mapbox</a>',    maxZoom: 18,    accessToken: 123456    }).addTo(mymap);    </script></body>
 </html>
